backend/requirements.txt
langchain
langgraph
langchain-groq
fastapi
uvicorn
pydantic
python-dotenv